refactor: reload가 아닌 다른 방법으로 값 초기화
