--- src/util/web/disabledButton.js
@@ -1,8 +1,0 @@
-const disableButton = (button) => {
-  if (!button) return;
-  button.disabled = true;
-  button.style.backgroundColor = '#ccc';
-  button.style.cursor = 'not-allowed';
-};
-
-export default disableButton;
